[arcane,std] Déplace les classes 'BasicWriter' et 'BasicGenericWriter' dans le fichier 'BasicWriter.cc'
